Feature push * Added support for Fancy Entry + RoboMX workaround * Added support for /notice & /gnotice * Added support for /opmsg * Added support for IPSend * Added support for newline escapes (#\n#) + access restrictions * Added support for /message (private messaging) with custom formats * Added support for /setformat

+++ src/chat-server/color-formatter.h
@@ -94,7 +94,7 @@ using namespace ng::utils;
 #include "available_colors.inc"
     };
 
-    for (unsigned int i = 0; i < sizeof available_colors; ++i) {
+    for (size_t i = 0; i < sizeof available_colors; ++i) {
         result += COL(available_colors[i]);
         result += std::format("#c{}#  ", (int)available_colors[i]);
         if (i % AVAILABLE_COLOR_COLUMNS == AVAILABLE_COLOR_COLUMNS - 1) {
